Reduce scope of do/catch blocks

+++ core-data-playground.playground/contents.swift
@@ -1,4 +1,4 @@
-//: ## A Swift 2.0 Introduction to Core Data
+//: ## A Swift Introduction to Core Data
 //:
 //: Let's create a basic Core Data model and populate it with two related entities. The two entities in our data model will be City and Neighborhood. There is a one-to-many relationship between cities and neighborhoods. A City has many Neighborhoods and each Neighborhood belongs to one City. Here's a diagram representing the model.
 //:
@@ -54,7 +54,6 @@ nameAttribute.attributeType = NSAttributeType.StringAttributeType
 nameAttribute.optional = false
 nameAttribute.indexed = false
 
-
 var stateAttribute = NSAttributeDescription()
 stateAttribute.name = GROAttribute.State
 stateAttribute.attributeType = NSAttributeType.StringAttributeType
@@ -97,9 +96,11 @@ model.entities = [cityEntity, neighborhoodEntity]
 //: Create a [Persistent Store Coordinator](https://developer.apple.com/library/ios/documentation/DataManagement/Devpedia-CoreData/persistentStoreCoordinator.html) (PSC) to communicate with the model we've declared. The coordinator is typically attached to an on disk SQL store with a URL. Because we're in a playground an [in memory store](https://developer.apple.com/library/mac/Documentation/Cocoa/Conceptual/CoreData/Articles/cdUsingPersistentStores.html) is used instead. When creating the PSC, you may include various options in the configuration dictionary, including specifying things like [migration policies](https://developer.apple.com/library/mac/documentation/cocoa/conceptual/CoreDataVersioning/Articles/vmInitiating.html); we can ignore these options in our simplistic stack.
 
 var persistentStoreCoordinator = NSPersistentStoreCoordinator(managedObjectModel:model)
+
 do {
     try persistentStoreCoordinator.addPersistentStoreWithType(NSInMemoryStoreType, configuration: nil, URL: nil, options: nil)
-}catch{
+}
+catch {
     print("error creating psc: \(error)")
 }
 
@@ -110,7 +111,7 @@ managedObjectContext.persistentStoreCoordinator = persistentStoreCoordinator
 
 //: The stack is now setup and ready to use. Create a new `City` entity and insert it into the context. This is done by calling `insertNewObjectForEntityForName` on `NSEntityDescription` and including the name of the entity and the context that should create it. This will return an `NSManagedObject` instance for our entity. We can then set attribute values on the entity using key paths. A typical app, with a more complex data model, may [create subclasses of `NSManagedObject`](http://stackoverflow.com/questions/7947458/why-exactly-would-one-subclass-nsmanagedobject) for specific entities and set attributes using instance variables instead of key paths.
 
-var city: AnyObject! = NSEntityDescription.insertNewObjectForEntityForName(GROEntity.City, inManagedObjectContext: managedObjectContext)
+var city = NSEntityDescription.insertNewObjectForEntityForName(GROEntity.City, inManagedObjectContext: managedObjectContext)
 
 city.setValue("Seattle", forKeyPath: GROAttribute.Name)
 city.setValue("Washington", forKeyPath: GROAttribute.State)
@@ -134,7 +135,7 @@ var neighborhoods = [[GROAttribute.Name:"Loyal Heights", GROAttribute.Population
     [GROAttribute.Name:"Belltown", GROAttribute.Population:7399]]
 
 for obj in neighborhoods {
-    var neighborhood: AnyObject! = NSEntityDescription.insertNewObjectForEntityForName(GROEntity.Neighborhood, inManagedObjectContext: managedObjectContext)
+    var neighborhood = NSEntityDescription.insertNewObjectForEntityForName(GROEntity.Neighborhood, inManagedObjectContext: managedObjectContext)
     neighborhood.setValue(obj[GROAttribute.Name], forKey: GROAttribute.Name)
     neighborhood.setValue(obj[GROAttribute.Population], forKey: GROAttribute.Population)
     
@@ -155,7 +156,8 @@ NSNotificationCenter.defaultCenter().addObserver(delegate, selector: "handleDidS
 //: Save the context so it's populated with the entities.
 do {
     try managedObjectContext.save()
-}catch {
+}
+catch {
     print("error saving context: \(error)")
 }
 
@@ -168,7 +170,8 @@ var results:[NSManagedObject] = []
 
 do {
     results = try managedObjectContext.executeFetchRequest(fetchRequest) as! [NSManagedObject]
-} catch {
+}
+catch {
     print("error executing fetch request: \(error)")
 }
 
@@ -177,47 +180,71 @@ results[0].valueForKey(GROAttribute.Name)
 results[1].valueForKey(GROAttribute.Name)
 
 //: Every managed object is assigned managed object id from the context. An [`NSManagedObjectID`](https://developer.apple.com/library/mac/documentation/cocoa/reference/CoreDataFramework/Classes/NSManagedObjectID_Class/Reference/NSManagedObjectID.html) is a unique id that can be used to reference managed objects across contexts. Consider the example below where `secondObject` is returned from another context by referencing a managed object id.
-do{
-    var moid = results[0].objectID
-    var firstObject = try managedObjectContext.existingObjectWithID(moid)
+
+var firstObject: NSManagedObject?
+var secondObject: NSManagedObject?
+
+let moid = results[0].objectID
+let secondContext = NSManagedObjectContext(concurrencyType: NSManagedObjectContextConcurrencyType.PrivateQueueConcurrencyType)
+secondContext.persistentStoreCoordinator = persistentStoreCoordinator
+
+do {
+    firstObject = try managedObjectContext.existingObjectWithID(moid)
+    secondObject = try secondContext.existingObjectWithID(moid)
+}
+catch {
+    print("error finding objects: \(error)")
+}
     
-    var secondContext = NSManagedObjectContext(concurrencyType: NSManagedObjectContextConcurrencyType.PrivateQueueConcurrencyType)
-    secondContext.persistentStoreCoordinator = persistentStoreCoordinator
-    
-    var secondObject = try secondContext.existingObjectWithID(moid)
-    
-    firstObject.valueForKey(GROAttribute.Name)
-    secondObject.valueForKey(GROAttribute.Name)
+firstObject?.valueForKey(GROAttribute.Name)
+secondObject?.valueForKey(GROAttribute.Name)
     
 //: Attributes and relationships on a managed object may be changed. When the context is saved the changes made to the managed objects are persisted. In the example below the `population` attribute of a `Neighborhood` entity is changed and the context to saved.
     
-    fetchRequest = NSFetchRequest(entityName: GROEntity.Neighborhood)
-    fetchRequest.predicate = NSPredicate(format: "name = %@", "Belltown")
-    
+fetchRequest = NSFetchRequest(entityName: GROEntity.Neighborhood)
+fetchRequest.predicate = NSPredicate(format: "name = %@", "Belltown")
+
+do {
     results = try managedObjectContext.executeFetchRequest(fetchRequest) as! [NSManagedObject]
-    var managedObject = results[0]
-    managedObject.valueForKey(GROAttribute.Name)
-    
-    managedObject.setValue(1000, forKey: GROAttribute.Population)
-    
+}
+catch {
+    print("error executing fetch: \(error)")
+}
+
+var managedObject = results.first
+
+managedObject?.valueForKey(GROAttribute.Name)
+managedObject?.setValue(1000, forKey: GROAttribute.Population)
+
+do {
+    let objectId = (managedObject?.objectID)!
     try managedObjectContext.save()
-    managedObject = try managedObjectContext.existingObjectWithID(managedObject.objectID)
-    
-    managedObject.valueForKey(GROAttribute.Name)
+    managedObject = try managedObjectContext.existingObjectWithID(objectId)
+}
+catch {
+    print("error finding object: \(error)")
+}
+
+managedObject?.valueForKey(GROAttribute.Name)
+
+if let managedObject = managedObject {
     
 //: Objects can be deleted through the context. Once deleted, they can no longer be retrieved by object id.
-    
-    managedObjectContext.deleteObject(managedObject)
-    try managedObjectContext.save()
-    
-//: should fail, object with id not found
     do {
-        managedObject = try managedObjectContext.existingObjectWithID(managedObject.objectID)
-    } catch {
-        print("error executing fetch request: \(error)")
+        managedObjectContext.deleteObject(managedObject)
+        try managedObjectContext.save()
+    }
+    catch {
+        print("error deleting object: \(error)")
     }
     
-}catch {
-    print(error)
+//: This should fail, object with id not found. The object has been deleted.
+    do {
+        let managedObject = try managedObjectContext.existingObjectWithID(managedObject.objectID)
+    }
+    catch {
+        print("error executing fetch request: \(error)")
+    }
 }
+
 //: That wraps up a basic introduction to Core Data using a Swift and a Playground. The Core Data framework is big and there's [much more explore](http://www.objc.io/issue-4/). For more information, consider reading through the [Core Data Programming Guide](https://developer.apple.com/library/mac/documentation/Cocoa/Conceptual/CoreData/Articles/cdBasics.html) or looking at the source for a Core Data [template project in Xcode](http://code.tutsplus.com/tutorials/core-data-from-scratch-core-data-stack--cms-20926).
